requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.3
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==6.0.2
django-dotenv==1.4.2
django-jazzmin==3.0.1
django-querycount==0.8.3
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.15
flower==2.0.1
humanize==4.15.0
idna==3.11
inflection==0.5.1
kombu==5.6.2
packaging==26.0
prometheus_client==0.25.0
prompt_toolkit==3.0.52
psycopg2==2.9.11
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2026.1.post1
PyYAML==6.0.3
redis==7.4.0
requests==2.31.0
six==1.17.0
sqlparse==0.5.5
tornado==6.5.5
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0